# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Personal Finance",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reminders": {
            "post": {
                "tags": [
                    "Reminders"
                ],
                "summary": "Create a new reminder",
                "operationId": "createReminder",
                "requestBody": {
                    "description": "Reminder data to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reminder"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reminder successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reminder added successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Reminder"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/monthly": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Generate Monthly Report",
                "operationId": "generateMonthlyReport",
                "requestBody": {
                    "description": "Monthly report generation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "month",
                                    "year"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "month": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2025
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Monthly report successfully generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MonthlyReportResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Reminder": {
                "required": [
                    "user_id",
                    "title",
                    "amount",
                    "due_date"
                ],
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Pay bills"
                    },
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-25"
                    }
                },
                "type": "object"
            },
            "MonthlyReportResponse": {
                "required": [
                    "month",
                    "total_income",
                    "total_expense",
                    "balance",
                    "categories"
                ],
                "properties": {
                    "month": {
                        "type": "string",
                        "example": "January 2025"
                    },
                    "total_income": {
                        "type": "number",
                        "format": "float",
                        "example": 1000
                    },
                    "total_expense": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "balance": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "category": {
                                    "type": "string",
                                    "example": "Food"
                                },
                                "amount": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 300
                                },
                                "percentage": {
                                    "type": "number",
                                    "format": "float",
                                    "example": 30
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}